+++ contracts/strategies/interfaces/IVectorFortress.sol
@@ -6,7 +6,7 @@ pragma solidity 0.8.21;
 /// @title Interface for FortiFi Vector Fortresses
 interface IVectorFortress {
     function deposit(uint amount, address user) external returns(uint);
-    function withdrawVector(address user, uint slippage) external;
+    function withdrawVector(address user, uint16 slippage) external;
     function recoverERC20(address to, address token, uint amount) external;
     function balanceOf(address holder) external view returns(uint256);
 }
